move target finding functions into reparent-utils
