+++ editor/src/components/canvas/canvas-strategies/strategies/reparent-utils.ts
@@ -1,9 +1,4 @@
 import type { ProjectContentTreeRoot } from '../../../assets'
-import {
-  addImport,
-  emptyImports,
-  mergeImports,
-} from '../../../../core/workers/common/project-file-utils'
 import type { AllElementProps } from '../../../editor/store/editor-state'
 import { withUnderlyingTarget } from '../../../editor/store/editor-state'
 import type { ElementPath, Imports, NodeModules } from '../../../../core/shared/project-file-types'
@@ -13,18 +8,9 @@ import type {
   ElementInstanceMetadataMap,
   JSXElementChild,
 } from '../../../../core/shared/element-template'
-import {
-  isIntrinsicElement,
-  isJSXElement,
-  JSXElement,
-  walkElement,
-} from '../../../../core/shared/element-template'
+import { isJSXConditionalExpression, isJSXElement } from '../../../../core/shared/element-template'
 import * as EP from '../../../../core/shared/element-path'
-import {
-  getImportsFor,
-  getRequiredImportsForElement,
-  importedFromWhere,
-} from '../../../editor/import-utils'
+import { getRequiredImportsForElement } from '../../../editor/import-utils'
 import { forceNotNull } from '../../../../core/shared/optional-utils'
 import { addImportsToFile } from '../../commands/add-imports-to-file-command'
 import type { BuiltInDependencies } from '../../../../core/es-modules/package-manager/built-in-dependencies-list'
@@ -34,7 +20,6 @@ import { getStoryboardElementPath } from '../../../../core/model/scene-utils'
 import { generateUidWithExistingComponents } from '../../../../core/model/element-template-utils'
 import { addElement } from '../../commands/add-element-command'
 import type { CustomStrategyState, InteractionCanvasState } from '../canvas-strategy-types'
-import { InteractionLifecycle } from '../canvas-strategy-types'
 import { duplicateElement } from '../../commands/duplicate-element-command'
 import { wildcardPatch } from '../../commands/wildcard-patch-command'
 import { hideInNavigatorCommand } from '../../commands/hide-in-navigator-command'
@@ -43,7 +28,7 @@ import type { InsertionPath } from '../../../editor/store/insertion-path'
 import {
   childInsertionPath,
   getElementPathFromInsertionPath,
-  isChildInsertionPath,
+  getInsertionPath,
 } from '../../../editor/store/insertion-path'
 import { getUtopiaID } from '../../../../core/shared/uid-utils'
 import type { IndexPosition } from '../../../../utils/utils'
@@ -51,7 +36,15 @@ import { fastForEach } from '../../../../core/shared/utils'
 import { addElements } from '../../commands/add-elements-command'
 import type { ElementPathTrees } from '../../../../core/shared/element-path-tree'
 import { getRequiredGroupTrueUps } from '../../commands/queue-group-true-up-command'
-import type { RemixRoutingTable } from '../../../editor/store/remix-derived-data'
+import type { Either } from '../../../../core/shared/either'
+import { left, right } from '../../../../core/shared/either'
+import { maybeBranchConditionalCase } from '../../../../core/model/conditionals'
+import type { NonEmptyArray } from '../../../../core/shared/array-utils'
+import { mapDropNulls, isNonEmptyArray } from '../../../../core/shared/array-utils'
+import type { MaybeInfinityCanvasRectangle } from '../../../../core/shared/math-utils'
+import { isInfinityRectangle } from '../../../../core/shared/math-utils'
+import { isElementRenderedBySameComponent } from './reparent-helpers/reparent-helpers'
+import type { ParsedCopyData } from '../../../../utils/clipboard'
 
 interface GetReparentOutcomeResult {
   commands: Array<CanvasCommand>
@@ -292,4 +285,282 @@ export function placeholderCloneCommands(
     }
   })
   return { commands: commands, duplicatedElementNewUids: duplicatedElementNewUids }
+}
+
+function rectangleSizesEqual(
+  a: MaybeInfinityCanvasRectangle | null,
+  b: MaybeInfinityCanvasRectangle | null,
+): boolean {
+  if (a == null || b == null || isInfinityRectangle(a) || isInfinityRectangle(b)) {
+    return false
+  }
+
+  return a.height === b.height && a.width === b.width
+}
+
+export type ReparentTargetForPaste =
+  | {
+      type: 'sibling'
+      siblingPath: ElementPath
+      parentPath: InsertionPath
+    }
+  | { type: 'parent'; parentPath: InsertionPath }
+
+type PasteParentNotFoundError =
+  | 'Cannot find a suitable parent'
+  | 'Cannot insert component instance into component definition'
+
+function checkComponentNotInsertedIntoOwnDefinition(
+  selectedViews: NonEmptyArray<ElementPath>,
+  metadata: ElementInstanceMetadataMap,
+  elementsToInsert: JSXElementChild[],
+): boolean {
+  const parentTarget = EP.getCommonParentOfNonemptyPathArray(selectedViews, true)
+
+  const jsxElements = elementsToInsert.filter(isJSXElement)
+
+  return jsxElements.some((element) =>
+    isElementRenderedBySameComponent(metadata, parentTarget, element),
+  )
+}
+
+function insertIntoSlot(
+  selectedViews: NonEmptyArray<ElementPath>,
+  metadata: ElementInstanceMetadataMap,
+  projectContents: ProjectContentTreeRoot,
+  elementPathTrees: ElementPathTrees,
+  numberOfElementsToInsert: number,
+): ReparentTargetForPaste | null {
+  const targetPath = selectedViews[0]
+  const parentPath = EP.parentPath(targetPath)
+  const parentElement = withUnderlyingTarget(parentPath, projectContents, null, (_, element) => {
+    return element
+  })
+
+  if (parentElement == null || !isJSXConditionalExpression(parentElement)) {
+    return null
+  }
+
+  const wrapperFragmentUID = generateUidWithExistingComponents(projectContents)
+  const conditionalCase = maybeBranchConditionalCase(parentPath, parentElement, targetPath)
+  if (conditionalCase == null) {
+    return null
+  }
+
+  const parentInsertionPath = getInsertionPath(
+    targetPath,
+    projectContents,
+    metadata,
+    elementPathTrees,
+    wrapperFragmentUID,
+    numberOfElementsToInsert,
+  )
+
+  if (parentInsertionPath == null) {
+    return null
+  }
+
+  return { type: 'parent', parentPath: parentInsertionPath }
+}
+
+function pasteNextToSameSizedElement(
+  copyData: ParsedCopyData,
+  selectedViews: NonEmptyArray<ElementPath>,
+  metadata: ElementInstanceMetadataMap,
+): ReparentTargetForPaste | null {
+  const targetPath = selectedViews[0]
+  const elementPasteEntry = copyData.elementPaste.at(0)
+  if (elementPasteEntry == null) {
+    return null
+  }
+
+  const selectedViewAABB = MetadataUtils.getFrameInCanvasCoords(targetPath, metadata)
+  // if the pasted item's BB is the same size as the selected item's BB
+  const pastedElementAABB = MetadataUtils.getFrameInCanvasCoords(
+    elementPasteEntry.originalElementPath,
+    copyData.originalContextMetadata,
+  )
+  // if the selected item's parent is autolayouted
+  const parentInstance = MetadataUtils.findElementByElementPath(metadata, EP.parentPath(targetPath))
+
+  const isSelectedViewParentAutolayouted = MetadataUtils.isFlexLayoutedContainer(parentInstance)
+
+  const pastingAbsoluteToAbsolute =
+    MetadataUtils.isPositionAbsolute(
+      MetadataUtils.findElementByElementPath(metadata, targetPath),
+    ) &&
+    MetadataUtils.isPositionAbsolute(
+      MetadataUtils.findElementByElementPath(
+        copyData.originalContextMetadata,
+        elementPasteEntry.originalElementPath,
+      ),
+    )
+
+  const pastedElementNames = mapDropNulls(
+    (element) => MetadataUtils.getJSXElementName(element.element),
+    copyData.elementPaste,
+  )
+
+  const parentPath = EP.parentPath(targetPath)
+  const targetElementSupportsInsertedElement = MetadataUtils.canInsertElementsToTargetText(
+    parentPath,
+    metadata,
+    pastedElementNames,
+  )
+
+  if (
+    rectangleSizesEqual(selectedViewAABB, pastedElementAABB) &&
+    (isSelectedViewParentAutolayouted || pastingAbsoluteToAbsolute) &&
+    targetElementSupportsInsertedElement
+  ) {
+    return {
+      type: 'sibling',
+      siblingPath: targetPath,
+      parentPath: childInsertionPath(EP.parentPath(targetPath)),
+    }
+  }
+  return null
+}
+
+function pasteIntoParentOrGrandparent(
+  elementsToInsert: JSXElementChild[],
+  projectContents: ProjectContentTreeRoot,
+  selectedViews: NonEmptyArray<ElementPath>,
+  metadata: ElementInstanceMetadataMap,
+  elementPathTree: ElementPathTrees,
+): ReparentTargetForPaste | null {
+  const pastedElementNames = mapDropNulls(
+    (element) => (element.type === 'JSX_ELEMENT' ? element.name : null),
+    elementsToInsert,
+  )
+
+  const parentTarget = EP.getCommonParentOfNonemptyPathArray(selectedViews, true)
+
+  // paste into parent
+  const targetElementSupportsInsertedElement = MetadataUtils.canInsertElementsToTargetText(
+    parentTarget,
+    metadata,
+    pastedElementNames,
+  )
+  if (
+    MetadataUtils.targetSupportsChildren(
+      projectContents,
+      metadata,
+      parentTarget,
+      elementPathTree,
+    ) &&
+    targetElementSupportsInsertedElement
+  ) {
+    return { type: 'parent', parentPath: childInsertionPath(parentTarget) }
+  }
+
+  // paste into parent of parent
+  const parentOfSelected = EP.parentPath(parentTarget)
+  if (
+    MetadataUtils.targetSupportsChildren(
+      projectContents,
+      metadata,
+      parentOfSelected,
+      elementPathTree,
+    )
+  ) {
+    return { type: 'parent', parentPath: childInsertionPath(parentOfSelected) }
+  }
+  return null
+}
+
+export function getTargetParentForOneShotInsertion(
+  storyboardPath: ElementPath,
+  projectContents: ProjectContentTreeRoot,
+  selectedViews: Array<ElementPath>,
+  metadata: ElementInstanceMetadataMap,
+  elementsToInsert: JSXElementChild[],
+  elementPathTree: ElementPathTrees,
+): Either<PasteParentNotFoundError, ReparentTargetForPaste> {
+  if (!isNonEmptyArray(selectedViews)) {
+    return right({ type: 'parent', parentPath: childInsertionPath(storyboardPath) })
+  }
+
+  if (checkComponentNotInsertedIntoOwnDefinition(selectedViews, metadata, elementsToInsert)) {
+    return left('Cannot insert component instance into component definition')
+  }
+
+  const insertIntoSlotResult = insertIntoSlot(
+    selectedViews,
+    metadata,
+    projectContents,
+    elementPathTree,
+    elementsToInsert.length,
+  )
+  if (insertIntoSlotResult != null) {
+    return right(insertIntoSlotResult)
+  }
+
+  const pasteIntoParentOrGrandparentResult = pasteIntoParentOrGrandparent(
+    elementsToInsert,
+    projectContents,
+    selectedViews,
+    metadata,
+    elementPathTree,
+  )
+  if (pasteIntoParentOrGrandparentResult != null) {
+    return right(pasteIntoParentOrGrandparentResult)
+  }
+  return left('Cannot find a suitable parent')
+}
+
+export function getTargetParentForPaste(
+  storyboardPath: ElementPath,
+  projectContents: ProjectContentTreeRoot,
+  selectedViews: Array<ElementPath>,
+  metadata: ElementInstanceMetadataMap,
+  copyData: ParsedCopyData,
+  elementPathTree: ElementPathTrees,
+): Either<PasteParentNotFoundError, ReparentTargetForPaste> {
+  if (!isNonEmptyArray(selectedViews)) {
+    return right({ type: 'parent', parentPath: childInsertionPath(storyboardPath) })
+  }
+  const pastedJSXElements = mapDropNulls(
+    (p) =>
+      MetadataUtils.getJSXElementFromMetadata(
+        copyData.originalContextMetadata,
+        p.originalElementPath,
+      ),
+    copyData.elementPaste,
+  )
+  if (checkComponentNotInsertedIntoOwnDefinition(selectedViews, metadata, pastedJSXElements)) {
+    return left('Cannot insert component instance into component definition')
+  }
+
+  const insertIntoSlotResult = insertIntoSlot(
+    selectedViews,
+    metadata,
+    projectContents,
+    elementPathTree,
+    copyData.elementPaste.length,
+  )
+  if (insertIntoSlotResult != null) {
+    return right(insertIntoSlotResult)
+  }
+
+  const pasteNextToSameSizedElementResult = pasteNextToSameSizedElement(
+    copyData,
+    selectedViews,
+    metadata,
+  )
+  if (pasteNextToSameSizedElementResult != null) {
+    return right(pasteNextToSameSizedElementResult)
+  }
+
+  const pasteIntoParentOrGrandparentResult = pasteIntoParentOrGrandparent(
+    copyData.elementPaste.map((e) => e.element),
+    projectContents,
+    selectedViews,
+    metadata,
+    elementPathTree,
+  )
+  if (pasteIntoParentOrGrandparentResult != null) {
+    return right(pasteIntoParentOrGrandparentResult)
+  }
+  return left('Cannot find a suitable parent')
 }
